services/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.9.1
motor==3.6.0
cryptography==43.0.1
boto3==1.35.0
python-multipart==0.0.9
defusedxml==0.7.1

# Logging and monitoring
structlog==24.4.0

# Development dependencies
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.2
